Move quotes tests to Servlet 60 bucket so it can repeat in 61
